flickr Monster tasks 1-6
It took me a little bit to figure out the quirks in storyboard mode and
segue'ing to a list of photos. I believe it's a bug in Xcode possibly.
I ended up redoing the storyboard portion of the photo list and issuing
clean to fix the issue I had. I had the same exact code and hook ups.